Simplify writing the screenshot to a file
Use a `io.Writer` on `*Screenshooter.Take` so that the pkg user can use
whatever they desire to write the screenshot. This also simplifies the
pkg since it no longer needs to handle file names and paths.

The CLI now defaults to writing to `./post.png` whenever the `-filename`
is not provided.