Add a command-line interface (#2)
* Move screenshot logic to its own package

Add a screenshooter package that implements taking screenshots of a
website using chromdp.

* Add a command-line interface

In order to use the Screenshooter, a CLI was added. It allows for
pointing to a URL and taking a screenshot of it.

* Improve Screenshooter's configuration

Provide a set of functions for configuring the currently available
parameters for the Screenshooter. This is a cleaner interface for
passing configuration to the object.

* Simplify writing the screenshot to a file

Use a `io.Writer` on `*Screenshooter.Take` so that the pkg user can use
whatever they desire to write the screenshot. This also simplifies the
pkg since it no longer needs to handle file names and paths.

The CLI now defaults to writing to `./post.png` whenever the `-filename`
is not provided.

* Fix the name of the Screenshotter

Proper English, you should use it.