Use chromedp.BrowserOption to simplify config
Previously, the config of the Screenshotter relied on
chromedp.ContextOption to define the browser interactions' behaviour.
This made it a bit hard to test, since the Screenshotter stored these
ContextOptions on its struct, and these are functions. In Go, the only
assertion we can make on functions is whether they're nil, otherwise a
test will always return that the comparision between two functions is
false.

On top of that, the context building process was complex, because it
depended on one of the configuration values of Screenshotter, timeout in
particular, to know whether it should rely on a context.Background or a
context.WithTimeout. This also made the timeout apply to the all
operation of taking and saving a screenshot, when in fact the only
concern is that it is not desired to wait eternally for a webpage to
respond. For that reason, Screenshotter now uses
chromedp.WithDialTimeout, which sets a timeout for communicating with
the browser instead.

A similar reasoning is behind using chromedp.WithBrowserDebugf instead
of chromedp.WithDebugf. This also returns a chromedp.BrowserOption
instead of an chromedp.ContextOptions.

The Screenshotter should now be much easier to test.